+++ doc/whats_new.rst
@@ -1,3 +1,23 @@
+0.4.1
+=====
+
+Enhancements
+------------
+
+    - :class:`nilearn.decomposition.CanICA` and
+      :class:`nilearn.decomposition.DictLearning` has new attribute
+      `components_img_` providing directly the components learned as
+      a Nifti image. This avoids the step of unmasking the attribute
+      `components_` which is true for older versions.
+
+Changes
+-------
+
+    - Using output attribute `components_` which is an extracted components
+      in :class:`nilearn.decomposition.CanICA` and
+      :class:`nilearn.decomposition.DictLearning` is deprecated and will
+      be removed in next two releases. Use `components_img_` instead.
+
 0.4.0
 =====